refactor(chat): isolate runtime workspace resolution
